refactor(support): move log export back to Support Center
#8128 moved the log-export tile into Developer Options on the argument
that regular users never need it. They do: it is the only way to hand
someone the full log file when Report a Bug's attached buffer is not
enough, and burying it behind the seven-tap developer unlock puts it out
of reach of the people who get asked for logs.

So the tile goes back where it was, between Request a Feature and the
account-deletion entry, ungated on auth exactly as before, restored to
Save Logs / "Export logs to file for manual sending" so all 22 locales
keep the translations #8128 deleted. devOptionsExportLogs and its
subtitle go, along with their _knownUntranslatedDebt entries; the six
outcome strings were already shared and are untouched.

What does not revert is the layering. #8128 routed the export through
ExportLogsCubit so the widget never sees BugReportService, and that
stays -- the tile is a Page/View split whose BlocProvider is keyed on
the watched services plus auth state, and all six LogExportStatus arms
move back verbatim. Copy Logs stays deleted, and so does the share
sheet's EXTRA_TEXT decoy that it existed to work around: every
Android share-sheet fix from #8128 is untouched.

The four #8113 / #8114 outcome tests move back into
support_center_screen_test.dart alongside the three #8128 added, now
verifying the export is called with SupportCenterScreen.
