+++ mobile/e2e/maestro/asserts/assertSettings.yaml
@@ -57,6 +57,10 @@ appId: co.openvine.app.staging
     ProofMode Info
     Learn about ProofMode verification and authenticity
 
+- assertVisible: |-
+    Save Logs
+    Export logs to file for manual sending
+
 - assertVisible: ADVANCED ACCOUNT OPTIONS
 
 - assertVisible: |-
