+++ mobile/lib/l10n/app_de.arb
@@ -1634,6 +1634,8 @@
     "supportReportBugSubtitle": "Technische Probleme mit der App",
     "supportRequestFeature": "Feature wünschen",
     "supportRequestFeatureSubtitle": "Eine Verbesserung oder ein neues Feature vorschlagen",
+    "supportSaveLogs": "Logs speichern",
+    "supportSaveLogsSubtitle": "Logs als Datei exportieren, um sie manuell zu senden",
     "supportFaq": "FAQ",
     "supportFaqSubtitle": "Häufige Fragen & Antworten",
     "supportFamily": "Divine Family",
