+++ mobile/lib/l10n/app_fil.arb
@@ -1134,6 +1134,8 @@
     "supportReportBugSubtitle": "Mga technical issue sa app",
     "supportRequestFeature": "Humiling ng Feature",
     "supportRequestFeatureSubtitle": "Magmungkahi ng improvement o bagong feature",
+    "supportSaveLogs": "I-save ang Logs",
+    "supportSaveLogsSubtitle": "I-export ang logs sa file para manual na maipadala",
     "supportFaq": "FAQ",
     "supportFaqSubtitle": "Mga karaniwang tanong at sagot",
     "supportFamily": "Divine Family",
