+++ mobile/lib/l10n/app_fr.arb
@@ -1634,6 +1634,8 @@
     "supportReportBugSubtitle": "Problèmes techniques avec l'app",
     "supportRequestFeature": "Demander une fonctionnalité",
     "supportRequestFeatureSubtitle": "Suggérer une amélioration ou une nouvelle fonctionnalité",
+    "supportSaveLogs": "Sauvegarder les logs",
+    "supportSaveLogsSubtitle": "Exporter les logs vers un fichier pour envoi manuel",
     "supportFaq": "FAQ",
     "supportFaqSubtitle": "Questions et réponses courantes",
     "supportFamily": "Divine Family",
