+++ mobile/lib/l10n/app_id.arb
@@ -1631,6 +1631,8 @@
     "supportReportBugSubtitle": "Masalah teknis dengan aplikasi",
     "supportRequestFeature": "Minta Fitur",
     "supportRequestFeatureSubtitle": "Sarankan perbaikan atau fitur baru",
+    "supportSaveLogs": "Simpan Log",
+    "supportSaveLogsSubtitle": "Ekspor log ke file untuk pengiriman manual",
     "supportFaq": "FAQ",
     "supportFaqSubtitle": "Pertanyaan & jawaban umum",
     "supportFamily": "Divine Family",
