+++ mobile/lib/l10n/app_ms.arb
@@ -2376,6 +2376,8 @@
   "supportReportBugSubtitle": "Isu teknikal dengan apl",
   "supportRequestFeature": "Minta Ciri Baharu",
   "supportRequestFeatureSubtitle": "Cadangkan penambahbaikan atau ciri baharu",
+  "supportSaveLogs": "Simpan Log",
+  "supportSaveLogsSubtitle": "Eksport log ke fail untuk penghantaran manual",
   "supportFaq": "Soalan Lazim",
   "supportFaqSubtitle": "Soalan & jawapan lazim",
   "supportFamily": "Divine Family",
