+++ mobile/lib/l10n/app_nl.arb
@@ -1634,6 +1634,8 @@
     "supportReportBugSubtitle": "Technische problemen met de app",
     "supportRequestFeature": "Functie aanvragen",
     "supportRequestFeatureSubtitle": "Stel een verbetering of nieuwe functie voor",
+    "supportSaveLogs": "Logs opslaan",
+    "supportSaveLogsSubtitle": "Exporteer logs naar bestand om handmatig te versturen",
     "supportFaq": "Veelgestelde vragen",
     "supportFaqSubtitle": "Veelgestelde vragen & antwoorden",
     "supportFamily": "Divine Family",
