+++ mobile/lib/l10n/app_pl.arb
@@ -1634,6 +1634,8 @@
     "supportReportBugSubtitle": "Problemy techniczne z aplikacją",
     "supportRequestFeature": "Poproś o funkcję",
     "supportRequestFeatureSubtitle": "Zasugeruj usprawnienie lub nową funkcję",
+    "supportSaveLogs": "Zapisz logi",
+    "supportSaveLogsSubtitle": "Eksportuj logi do pliku do ręcznego wysłania",
     "supportFaq": "FAQ",
     "supportFaqSubtitle": "Częste pytania i odpowiedzi",
     "supportFamily": "Divine Family",
