+++ mobile/lib/l10n/app_pt.arb
@@ -1634,6 +1634,8 @@
     "supportReportBugSubtitle": "Problemas técnicos com o app",
     "supportRequestFeature": "Pedir uma funcionalidade",
     "supportRequestFeatureSubtitle": "Sugira uma melhoria ou nova funcionalidade",
+    "supportSaveLogs": "Salvar logs",
+    "supportSaveLogsSubtitle": "Exportar logs para um arquivo para envio manual",
     "supportFaq": "FAQ",
     "supportFaqSubtitle": "Perguntas e respostas comuns",
     "supportFamily": "Divine Family",
