+++ mobile/lib/l10n/app_tr.arb
@@ -1631,6 +1631,8 @@
     "supportReportBugSubtitle": "Uygulamayla ilgili teknik sorunlar",
     "supportRequestFeature": "Özellik İste",
     "supportRequestFeatureSubtitle": "Bir iyileştirme veya yeni özellik öner",
+    "supportSaveLogs": "Günlükleri Kaydet",
+    "supportSaveLogsSubtitle": "Manuel göndermek için günlükleri dosyaya aktar",
     "supportFaq": "SSS",
     "supportFaqSubtitle": "Sık sorulan sorular ve cevapları",
     "supportFamily": "Divine Family",
