+++ mobile/lib/l10n/app_ur.arb
@@ -2376,6 +2376,8 @@
   "supportReportBugSubtitle": "ایپ کے تکنیکی مسائل",
   "supportRequestFeature": "فیچر کی درخواست کریں",
   "supportRequestFeatureSubtitle": "کوئی بہتری یا نیا فیچر تجویز کریں",
+  "supportSaveLogs": "لاگز محفوظ کریں",
+  "supportSaveLogsSubtitle": "دستی بھیجنے کے لیے لاگز فائل میں ایکسپورٹ کریں",
   "supportFaq": "عمومی سوالات",
   "supportFaqSubtitle": "عام سوالات اور جوابات",
   "supportFamily": "Divine Family",
