+++ mobile/lib/l10n/app_zh.arb
@@ -2376,6 +2376,8 @@
   "supportReportBugSubtitle": "应用的技术问题",
   "supportRequestFeature": "功能建议",
   "supportRequestFeatureSubtitle": "提出改进建议或新功能想法",
+  "supportSaveLogs": "保存日志",
+  "supportSaveLogsSubtitle": "导出日志到文件，手动发送",
   "supportFaq": "常见问题",
   "supportFaqSubtitle": "常见问题与解答",
   "supportFamily": "Divine Family",
