+++ mobile/lib/screens/settings/support_center_screen.dart
@@ -3,8 +3,10 @@
 
 import 'package:divine_ui/divine_ui.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
+import 'package:openvine/blocs/export_logs/export_logs_cubit.dart';
 import 'package:openvine/constants/app_constants.dart';
 import 'package:openvine/extensions/safe_pop_extension.dart';
 import 'package:openvine/l10n/l10n.dart';
@@ -74,6 +76,7 @@ class SupportCenterScreen extends ConsumerWidget {
                   subtitle: l10n.supportRequestFeatureSubtitle,
                   onTap: () => _showFeatureRequest(context),
                 ),
+              const _ExportLogsTile(),
               // #6335 was filed from here by someone who could not find
               // deletion, so it has to be reachable from here too. It sits
               // above the outbound links rather than last so it stays on
@@ -270,6 +273,124 @@ class _SupportTile extends StatelessWidget {
         color: trailingColor ?? context.vineColors.mutedText,
       ),
       onTap: onTap,
+    );
+  }
+}
+
+/// Hands the captured log file to the system share sheet.
+///
+/// Goes through [ExportLogsCubit] rather than calling `BugReportService`
+/// from here, so the widget layer never sees the service type.
+class _ExportLogsTile extends ConsumerWidget {
+  const _ExportLogsTile();
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final bugReportService = ref.watch(bugReportServiceProvider);
+    final authService = ref.watch(authServiceProvider);
+    final authState = ref.watch(currentAuthStateProvider);
+    return BlocProvider<ExportLogsCubit>(
+      key: ValueKey((bugReportService, authService, authState)),
+      create: (_) => ExportLogsCubit(
+        bugReportService: bugReportService,
+        currentScreen: 'SupportCenterScreen',
+        userPubkey: authService.currentPublicKeyHex,
+      ),
+      child: const _ExportLogsTileView(),
+    );
+  }
+}
+
+class _ExportLogsTileView extends StatelessWidget {
+  const _ExportLogsTileView();
+
+  @override
+  Widget build(BuildContext context) {
+    final l10n = context.l10n;
+    return BlocListener<ExportLogsCubit, ExportLogsState>(
+      listenWhen: (previous, current) => previous.status != current.status,
+      listener: _showOutcome,
+      child: _SupportTile(
+        icon: DivineIconName.save,
+        title: l10n.supportSaveLogs,
+        subtitle: l10n.supportSaveLogsSubtitle,
+        // Resolve the popover anchor before the cubit's first await — the
+        // iPad idiom (including iOS builds on Apple Silicon Macs) rejects the
+        // share sheet without it.
+        onTap: () => context.read<ExportLogsCubit>().export(
+          sharePositionOrigin: shareAnchorForContext(context),
+        ),
+      ),
+    );
+  }
+
+  static void _showOutcome(BuildContext context, ExportLogsState state) {
+    final l10n = context.l10n;
+    final messenger = ScaffoldMessenger.of(context);
+
+    if (state.status == ExportLogsStatus.exporting) {
+      messenger.showSnackBar(
+        DivineSnackbarContainer.snackBar(
+          l10n.supportExportingLogs,
+          duration: const Duration(seconds: 2),
+        ),
+      );
+      return;
+    }
+
+    messenger.hideCurrentSnackBar();
+
+    switch (state.status) {
+      case ExportLogsStatus.idle:
+      case ExportLogsStatus.exporting:
+      case ExportLogsStatus.cancelled:
+        // Cancelled means the user backed out — they know what they did.
+        return;
+      case ExportLogsStatus.noLogs:
+        messenger.showSnackBar(
+          DivineSnackbarContainer.snackBar(
+            l10n.supportNoLogsToExport,
+            duration: const Duration(seconds: 8),
+          ),
+        );
+        return;
+      case ExportLogsStatus.unconfirmed:
+        messenger.showSnackBar(
+          DivineSnackbarContainer.snackBar(
+            l10n.supportExportLogsUnconfirmed,
+            duration: const Duration(seconds: 6),
+          ),
+        );
+        return;
+      case ExportLogsStatus.failed:
+        messenger.showSnackBar(
+          DivineSnackbarContainer.snackBar(
+            l10n.supportExportLogsFailed,
+            error: true,
+          ),
+        );
+        return;
+      case ExportLogsStatus.shared:
+      case ExportLogsStatus.saved:
+        break;
+    }
+
+    final filePath = state.filePath;
+    if (filePath == null) return;
+
+    final cubit = context.read<ExportLogsCubit>();
+    messenger.showSnackBar(
+      DivineSnackbarContainer.snackBar(
+        l10n.supportLogsSavedTo(filePath),
+        duration: const Duration(seconds: 8),
+        actionLabel: l10n.supportRevealLogsAction,
+        onActionPressed: () {
+          // DivineSnackbarContainer's action is a plain button, so unlike
+          // SnackBarAction it does not dismiss the banner for us.
+          messenger.hideCurrentSnackBar();
+          cubit.revealFile(filePath);
+        },
+      ),
     );
   }
 }
