+++ mobile/test/l10n/arb_consistency_test.dart
@@ -484,12 +484,9 @@ void main() {
 // Keys intentionally allowed to fall back to English until a translation pass.
 // Keep this list small and reviewable so new translation gaps stay visible.
 const _knownUntranslatedDebt = <String>{
-  // Log-export copy (#8112 / #8113 / #8114 / #8127). Left in English until a
-  // human translation pass; machine-translating a diagnostic instruction the
-  // user has to follow exactly is how it stops meaning what it says. The
-  // devOptions pair is developer-only surface behind the seven-tap unlock.
-  'devOptionsExportLogs',
-  'devOptionsExportLogsSubtitle',
+  // Log-export copy (#8112 / #8113 / #8114). Left in English until a human
+  // translation pass; machine-translating a diagnostic instruction the user
+  // has to follow exactly is how it stops meaning what it says.
   'supportNoLogsToExport',
   'supportExportLogsUnconfirmed',
   // Account-enforcement translation remains tracked in #7765. The policy copy
