+++ mobile/test/screens/developer_options_screen_test.dart
@@ -96,9 +96,7 @@ void main() {
   group('renders', () {
     testWidgets(
       'renders the available Shorebird section through screen wiring',
-      (
-        tester,
-      ) async {
+      (tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
         var factoryCalls = 0;
         final updater = _AvailableShorebirdUpdater();
@@ -119,32 +117,6 @@ void main() {
         expect(find.text(l10n.devOptionsShorebirdNotChecked), findsOneWidget);
       },
     );
-
-    // #8127 moved log export here from Support Center, which regular users
-    // see. It belongs between the environment picker and Shorebird patches.
-    testWidgets('places Export Logs between environments and Shorebird', (
-      tester,
-    ) async {
-      final l10n = lookupAppLocalizations(const Locale('en'));
-      await pumpScreen(tester, await mockPrefs());
-
-      expect(find.text(l10n.devOptionsExportLogs), findsOneWidget);
-
-      final production = tester.getTopLeft(find.text('Production')).dy;
-      final exportLogs = tester
-          .getTopLeft(
-            find.text(l10n.devOptionsExportLogs),
-          )
-          .dy;
-      final shorebird = tester
-          .getTopLeft(
-            find.text(l10n.devOptionsShorebirdTitle),
-          )
-          .dy;
-
-      expect(exportLogs, greaterThan(production));
-      expect(exportLogs, lessThan(shorebird));
-    });
 
     testWidgets(
       'DeveloperOptionsScreen constrains menu content width on wide screens',
