+++ mobile/test/screens/settings/support_center_screen_test.dart
@@ -1,5 +1,7 @@
-// ABOUTME: Tests the Support Center resource links and account-deletion auth gate
-// ABOUTME: Regression cover for #6335, which was reported from this screen
+// ABOUTME: Tests the Support Center resource links, log export, and deletion gate
+// ABOUTME: Regression cover for #6335, #8113 and #8114, all reported from here
+
+import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -11,6 +13,7 @@ import 'package:openvine/providers/app_providers.dart';
 import 'package:openvine/screens/settings/support_center_screen.dart';
 import 'package:openvine/services/account_deletion_service.dart';
 import 'package:openvine/services/auth_service.dart';
+import 'package:openvine/services/bug_report_service.dart';
 import 'package:openvine/services/support_email_composer.dart';
 import 'package:openvine/services/zendesk_support_service.dart';
 import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
@@ -18,6 +21,8 @@ import 'package:url_launcher_platform_interface/url_launcher_platform_interface.
 import '../../helpers/url_launcher_test_double.dart';
 
 class _MockAuthService extends Mock implements AuthService {}
+
+class _MockBugReportService extends Mock implements BugReportService {}
 
 class _MockAccountDeletionService extends Mock
     implements AccountDeletionService {}
@@ -28,12 +33,14 @@ const _pubkeyHex =
 void main() {
   group(SupportCenterScreen, () {
     late _MockAuthService authService;
+    late _MockBugReportService bugReportService;
     late _MockAccountDeletionService accountDeletionService;
     final en = lookupAppLocalizations(const Locale('en'));
 
     setUp(() {
       ZendeskSupportService.resetForTesting();
       authService = _MockAuthService();
+      bugReportService = _MockBugReportService();
       accountDeletionService = _MockAccountDeletionService();
       when(() => authService.currentPublicKeyHex).thenReturn(_pubkeyHex);
     });
@@ -52,6 +59,7 @@ void main() {
           overrides: [
             authServiceProvider.overrideWithValue(authService),
             currentAuthStateProvider.overrideWith((ref) => authState),
+            bugReportServiceProvider.overrideWithValue(bugReportService),
             accountDeletionServiceProvider.overrideWithValue(
               accountDeletionService,
             ),
@@ -82,6 +90,140 @@ void main() {
       await tester.drag(find.byType(ListView), const Offset(0, -2000));
       await tester.pumpAndSettle();
     }
+
+    group('save logs', () {
+      Future<void> tapSaveLogs(
+        WidgetTester tester,
+        LogExportResult result,
+      ) async {
+        when(
+          () => bugReportService.exportLogsToFile(
+            currentScreen: any(named: 'currentScreen'),
+            userPubkey: any(named: 'userPubkey'),
+            sharePositionOrigin: any(named: 'sharePositionOrigin'),
+          ),
+        ).thenAnswer((_) async => result);
+
+        await pump(tester);
+        await tester.tap(find.text(en.supportSaveLogs));
+        await tester.pumpAndSettle();
+      }
+
+      testWidgets('reads its title and subtitle from l10n', (tester) async {
+        await pump(tester);
+
+        expect(find.text(en.supportSaveLogs), findsOneWidget);
+        expect(find.text(en.supportSaveLogsSubtitle), findsOneWidget);
+      });
+
+      testWidgets('exports with the signed-in pubkey and this screen', (
+        tester,
+      ) async {
+        await tapSaveLogs(tester, const LogExportResult.shared());
+
+        verify(
+          () => bugReportService.exportLogsToFile(
+            currentScreen: 'SupportCenterScreen',
+            userPubkey: _pubkeyHex,
+            sharePositionOrigin: any(named: 'sharePositionOrigin'),
+          ),
+        ).called(1);
+      });
+
+      testWidgets('ignores a second tap while an export is in flight', (
+        tester,
+      ) async {
+        final gate = Completer<LogExportResult>();
+        when(
+          () => bugReportService.exportLogsToFile(
+            currentScreen: any(named: 'currentScreen'),
+            userPubkey: any(named: 'userPubkey'),
+            sharePositionOrigin: any(named: 'sharePositionOrigin'),
+          ),
+        ).thenAnswer((_) => gate.future);
+
+        await pump(tester);
+        await tester.tap(find.text(en.supportSaveLogs));
+        await tester.pump();
+        await tester.tap(find.text(en.supportSaveLogs));
+        await tester.pump();
+
+        expect(find.text(en.supportExportingLogs), findsOneWidget);
+
+        gate.complete(const LogExportResult.shared());
+        await tester.pumpAndSettle();
+
+        verify(
+          () => bugReportService.exportLogsToFile(
+            currentScreen: any(named: 'currentScreen'),
+            userPubkey: any(named: 'userPubkey'),
+            sharePositionOrigin: any(named: 'sharePositionOrigin'),
+          ),
+        ).called(1);
+      });
+
+      // #8113: share_plus returns `unavailable` on Android whenever it cannot
+      // attach to an Activity, even though the sheet opened and the share may
+      // well have completed. Reporting that as a failure is what told the
+      // user log export was broken.
+      testWidgets('does not report failure when the outcome is unknown', (
+        tester,
+      ) async {
+        await tapSaveLogs(tester, const LogExportResult.unconfirmed());
+
+        expect(find.text(en.supportExportLogsFailed), findsNothing);
+        expect(find.text(en.supportExportLogsUnconfirmed), findsOneWidget);
+      });
+
+      testWidgets('stays silent when the user backs out', (tester) async {
+        await tapSaveLogs(tester, const LogExportResult.cancelled());
+
+        expect(find.text(en.supportExportLogsFailed), findsNothing);
+        expect(find.text(en.supportExportLogsUnconfirmed), findsNothing);
+        expect(find.text(en.supportNoLogsToExport), findsNothing);
+      });
+
+      // #8114: the buffer is memory-only, so a crash or force-quit empties
+      // it — exactly when the user needs to be told to reproduce first
+      // rather than conclude export is broken.
+      testWidgets('explains an empty buffer rather than failing', (
+        tester,
+      ) async {
+        await tapSaveLogs(tester, const LogExportResult.noLogs());
+
+        expect(find.text(en.supportNoLogsToExport), findsOneWidget);
+        expect(find.text(en.supportExportLogsFailed), findsNothing);
+      });
+
+      testWidgets('still reports a real failure', (tester) async {
+        await tapSaveLogs(tester, const LogExportResult.failed());
+
+        expect(find.text(en.supportExportLogsFailed), findsOneWidget);
+      });
+
+      testWidgets('offers to reveal a saved file', (tester) async {
+        when(
+          () => bugReportService.revealExportedFile(any()),
+        ).thenAnswer((_) async {});
+
+        await tapSaveLogs(
+          tester,
+          const LogExportResult.saved('/tmp/divine_logs.txt'),
+        );
+
+        expect(
+          find.text(en.supportLogsSavedTo('/tmp/divine_logs.txt')),
+          findsOneWidget,
+        );
+
+        await tester.tap(find.text(en.supportRevealLogsAction));
+        await tester.pumpAndSettle();
+
+        verify(
+          () => bugReportService.revealExportedFile('/tmp/divine_logs.txt'),
+        ).called(1);
+      });
+    });
 
     // #6335 was filed from this screen by a user who could not find how to
     // delete their account. Deletion has to be reachable from here.
